add torrent status and edit progress unit

+++ account/models.py
@@ -8,7 +8,7 @@ class Account(models.Model):
 class TorrentEntries(models.Model):
 	name=models.CharField(max_length=100, null=True)
 	hash_value=models.CharField(max_length=40, null=False)
-	progress=models.FloatField(default=0, null=True)
+	progress=models.PositiveSmallIntegerField(default=0, null=True)
 	download_rate=models.FloatField(default=0, null=True)
 	file_size=models.BigIntegerField(default=0, null=True)
 	downloaded_size=models.BigIntegerField(default=0, null=True)
